threaded video input
