+++ rio-server/package.json
@@ -29,6 +29,6 @@
     "threads": "^0.7.2",
     "twit": "^2.2.5",
     "ws": "^1.1.1",
-    "youtube-dl": "^1.11.1"
+    "youtube-dl": "1.11.0"
   }
 }
